backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML libraries
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
fer==22.5.1
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3

# Audio processing
librosa==0.10.1
soundfile==0.12.1

# Google Drive API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
